Add experimental export for minimal import
