+++ packages/core/src/data-editor-base.ts
@@ -1,5 +1,3 @@
-export type { OverlayImageEditorProps } from "./data-grid-overlay-editor/private/image-overlay-editor";
-export type { MarkdownDivProps } from "./markdown-div/markdown-div";
 export type { SpriteMap, HeaderIcon, Sprite } from "./data-grid/data-grid-sprites";
 export type { SpriteProps } from "./common/utils";
 export type { Theme } from "./common/styles";
@@ -8,12 +6,8 @@ export type { CustomRenderer } from "./data-grid/cells/cell-types";
 export type { SelectionBlending } from "./data-grid/use-selection-behavior";
 export type { GetRowThemeCallback } from "./data-grid/data-grid-render";
 
-export { DataEditorRef } from "./data-editor/data-editor";
-export { DataEditorAll as DataEditor, DataEditorAllProps as DataEditorProps } from "./data-editor-all";
+export * from "./data-editor/data-editor";
 export * from "./data-grid/data-grid-types";
-export { ImageOverlayEditor } from "./data-grid-overlay-editor/private/image-overlay-editor";
-export { default as MarkdownDiv } from "./markdown-div/markdown-div";
-export { GrowingEntry as TextCellEntry } from "./growing-entry/growing-entry";
 export { parseToRgba, withAlpha, blend, interpolateColors } from "./data-grid/color-parser";
 export {
     measureTextCached,
@@ -23,9 +17,3 @@ export {
 export { getDataEditorTheme as getDefaultTheme, useTheme } from "./common/styles";
 export { useColumnSizer } from "./data-editor/use-column-sizer";
 export { useCustomCells } from "./data-editor/use-custom-cells";
-
-/**
- * @category DataEditor
- * @hidden
- */
-export { DataEditorAll as default } from "./data-editor-all";
