+++ packages/core/src/data-grid/cells/image-cell.tsx
@@ -1,15 +1,14 @@
 /* eslint-disable react/display-name */
 import * as React from "react";
 import { ImageOverlayEditor } from "../../data-grid-overlay-editor/private/image-overlay-editor";
-import { drawImage } from "../data-grid-lib";
-import { GridCellKind, ImageCell } from "../data-grid-types";
-import type { InternalCellRenderer } from "./cell-types";
+import { roundedRect } from "../data-grid-drawing";
+import { BaseGridCell, GridCellKind, ImageCell } from "../data-grid-types";
+import type { BaseDrawArgs, InternalCellRenderer } from "./cell-types";
 
 export const imageCellRenderer: InternalCellRenderer<ImageCell> = {
     getAccessibilityString: c => c.data.join(", "),
     kind: GridCellKind.Image,
     needsHover: false,
-    useLabel: false,
     needsHoverPosition: false,
     draw: a => drawImage(a, a.cell.displayData ?? a.cell.data, a.cell.rounding, a.cell.contentAlign),
     measure: (_ctx, cell) => cell.data.length * 50,
@@ -57,3 +56,54 @@ export const imageCellRenderer: InternalCellRenderer<ImageCell> = {
         };
     },
 };
+
+const itemMargin = 4;
+
+function drawImage(
+    args: BaseDrawArgs,
+    data: readonly string[],
+    rounding: number = 4,
+    contentAlign?: BaseGridCell["contentAlign"]
+) {
+    const { rect, col, row, theme, ctx, imageLoader } = args;
+    const { x, y, height: h, width: w } = rect;
+
+    const imgHeight = h - theme.cellVerticalPadding * 2;
+    const images: (HTMLImageElement | ImageBitmap)[] = [];
+    let totalWidth = 0;
+    // eslint-disable-next-line unicorn/no-for-loop
+    for (let index = 0; index < data.length; index++) {
+        const i = data[index];
+        if (i.length === 0) continue;
+        const img = imageLoader.loadOrGetImage(i, col, row);
+
+        if (img !== undefined) {
+            images[index] = img;
+            const imgWidth = img.width * (imgHeight / img.height);
+            totalWidth += imgWidth + itemMargin;
+        }
+    }
+
+    if (totalWidth === 0) return;
+    totalWidth -= itemMargin;
+
+    let drawX = x + theme.cellHorizontalPadding;
+    if (contentAlign === "right") drawX = Math.floor(x + w - theme.cellHorizontalPadding - totalWidth);
+    else if (contentAlign === "center") drawX = Math.floor(x + w / 2 - totalWidth / 2);
+
+    for (const img of images) {
+        if (img === undefined) continue; //array is sparse
+        const imgWidth = img.width * (imgHeight / img.height);
+        if (rounding > 0) {
+            roundedRect(ctx, drawX, y + theme.cellVerticalPadding, imgWidth, imgHeight, rounding);
+            ctx.save();
+            ctx.clip();
+        }
+        ctx.drawImage(img, drawX, y + theme.cellVerticalPadding, imgWidth, imgHeight);
+        if (rounding > 0) {
+            ctx.restore();
+        }
+
+        drawX += imgWidth + itemMargin;
+    }
+}
