+++ packages/core/src/data-grid/data-grid-lib.ts
@@ -1,20 +1,9 @@
 /* eslint-disable unicorn/no-for-loop */
 import type { Theme } from "../common/styles";
-import {
-    DrilldownCellData,
-    Item,
-    GridSelection,
-    InnerGridCell,
-    SizedGridColumn,
-    Rectangle,
-    BaseGridCell,
-    BooleanEmpty,
-    BooleanIndeterminate,
-} from "./data-grid-types";
-import { degreesToRadians, direction } from "../common/utils";
+import type { Item, GridSelection, InnerGridCell, SizedGridColumn, Rectangle, BaseGridCell } from "./data-grid-types";
+import { direction } from "../common/utils";
 import React from "react";
 import type { BaseDrawArgs, PrepResult } from "./cells/cell-types";
-import { assertNever } from "../common/support";
 import { split as splitText, clearCache } from "canvas-hypertxt";
 
 export interface MappedGridColumn extends SizedGridColumn {
@@ -393,7 +382,7 @@ function drawSingleTextLine(
     }
 }
 
-function getEmHeight(ctx: CanvasRenderingContext2D, fontStyle: string): number {
+export function getEmHeight(ctx: CanvasRenderingContext2D, fontStyle: string): number {
     const textMetrics = measureTextCached("ABCi09jgqpy", ctx, fontStyle); // do not question the magic string
     return textMetrics.actualBoundingBoxAscent + textMetrics.actualBoundingBoxDescent;
 }
@@ -487,623 +476,6 @@ export function drawTextCell(
         if (isRtl) {
             ctx.direction = "inherit";
         }
-    }
-}
-
-export function drawNewRowCell(args: BaseDrawArgs, data: string, icon?: string) {
-    const { ctx, rect, hoverAmount, theme, spriteManager } = args;
-    const { x, y, width: w, height: h } = rect;
-    ctx.beginPath();
-    ctx.globalAlpha = hoverAmount;
-    ctx.rect(x, y, w, h);
-    ctx.fillStyle = theme.bgHeaderHovered;
-    ctx.fill();
-    ctx.globalAlpha = 1;
-    ctx.beginPath();
-
-    const alwaysShowIcon = data !== "";
-
-    let textX = 0;
-
-    if (icon !== undefined) {
-        const padding = 8;
-        const size = h - padding;
-        const px = x + padding / 2;
-        const py = y + padding / 2;
-
-        spriteManager.drawSprite(icon, "normal", ctx, px, py, size, theme, alwaysShowIcon ? 1 : hoverAmount);
-        textX = size;
-    } else {
-        textX = 24;
-        const finalLineSize = 12;
-        const lineSize = alwaysShowIcon ? finalLineSize : hoverAmount * finalLineSize;
-        const xTranslate = alwaysShowIcon ? 0 : (1 - hoverAmount) * finalLineSize * 0.5;
-
-        const padPlus = theme.cellHorizontalPadding + 4;
-        if (lineSize > 0) {
-            ctx.moveTo(x + padPlus + xTranslate, y + h / 2);
-            ctx.lineTo(x + padPlus + xTranslate + lineSize, y + h / 2);
-            ctx.moveTo(x + padPlus + xTranslate + lineSize * 0.5, y + h / 2 - lineSize * 0.5);
-            ctx.lineTo(x + padPlus + xTranslate + lineSize * 0.5, y + h / 2 + lineSize * 0.5);
-            ctx.lineWidth = 2;
-            ctx.strokeStyle = theme.bgIconHeader;
-            ctx.lineCap = "round";
-            ctx.stroke();
-        }
-    }
-
-    ctx.fillStyle = theme.textMedium;
-    ctx.fillText(data, textX + x + theme.cellHorizontalPadding + 0.5, y + h / 2 + getMiddleCenterBias(ctx, theme));
-    ctx.beginPath();
-}
-
-export function drawCheckbox(
-    ctx: CanvasRenderingContext2D,
-    theme: Theme,
-    checked: boolean | BooleanEmpty | BooleanIndeterminate,
-    x: number,
-    y: number,
-    width: number,
-    height: number,
-    highlighted: boolean,
-    hoverX: number = -20,
-    hoverY: number = -20,
-    maxSize: number = 32
-) {
-    const centerX = Math.floor(x + width / 2);
-    const centerY = Math.floor(y + height / 2);
-
-    const checkBoxWidth = Math.min(maxSize, height - theme.cellVerticalPadding * 2);
-
-    const hoverHelper = checkBoxWidth / 2;
-    const hovered = Math.abs(hoverX - width / 2) < hoverHelper && Math.abs(hoverY - height / 2) < hoverHelper;
-
-    const rectBordRadius = 4;
-    const posHelperChecked = checkBoxWidth / 2;
-
-    switch (checked) {
-        case true: {
-            ctx.beginPath();
-            roundedRect(
-                ctx,
-                centerX - checkBoxWidth / 2,
-                centerY - checkBoxWidth / 2,
-                checkBoxWidth,
-                checkBoxWidth,
-                rectBordRadius
-            );
-
-            ctx.fillStyle = highlighted ? theme.accentColor : theme.textMedium;
-            ctx.fill();
-
-            ctx.beginPath();
-            ctx.moveTo(
-                centerX - posHelperChecked + checkBoxWidth / 4.23,
-                centerY - posHelperChecked + checkBoxWidth / 1.97
-            );
-            ctx.lineTo(
-                centerX - posHelperChecked + checkBoxWidth / 2.42,
-                centerY - posHelperChecked + checkBoxWidth / 1.44
-            );
-            ctx.lineTo(
-                centerX - posHelperChecked + checkBoxWidth / 1.29,
-                centerY - posHelperChecked + checkBoxWidth / 3.25
-            );
-
-            ctx.strokeStyle = theme.bgCell;
-            ctx.lineJoin = "round";
-            ctx.lineCap = "round";
-            ctx.lineWidth = 1.9;
-            ctx.stroke();
-            break;
-        }
-
-        case BooleanEmpty:
-        case false: {
-            ctx.beginPath();
-            roundedRect(
-                ctx,
-                centerX - checkBoxWidth / 2 + 0.5,
-                centerY - checkBoxWidth / 2 + 0.5,
-                checkBoxWidth - 1,
-                checkBoxWidth - 1,
-                rectBordRadius
-            );
-
-            ctx.lineWidth = 1;
-            ctx.strokeStyle = hovered ? theme.textDark : theme.textMedium;
-            ctx.stroke();
-            break;
-        }
-
-        case BooleanIndeterminate: {
-            ctx.beginPath();
-            roundedRect(
-                ctx,
-                centerX - checkBoxWidth / 2,
-                centerY - checkBoxWidth / 2,
-                checkBoxWidth,
-                checkBoxWidth,
-                rectBordRadius
-            );
-
-            ctx.fillStyle = hovered ? theme.textMedium : theme.textLight;
-            ctx.fill();
-
-            ctx.beginPath();
-            ctx.moveTo(centerX - checkBoxWidth / 3, centerY);
-            ctx.lineTo(centerX + checkBoxWidth / 3, centerY);
-            ctx.strokeStyle = theme.bgCell;
-            ctx.lineCap = "round";
-            ctx.lineWidth = 1.9;
-            ctx.stroke();
-            break;
-        }
-
-        default:
-            assertNever(checked);
-    }
-}
-
-export function prepMarkerRowCell(args: BaseDrawArgs, lastPrep: PrepResult | undefined): Partial<PrepResult> {
-    const { ctx, theme } = args;
-    const newFont = `9px ${theme.fontFamily}`;
-    const result: Partial<PrepResult> = lastPrep ?? {};
-    if (result?.font !== newFont) {
-        ctx.font = newFont;
-        result.font = newFont;
-    }
-    result.deprep = deprepMarkerRowCell;
-    ctx.textAlign = "center";
-    return result;
-}
-
-export function deprepMarkerRowCell(args: Pick<BaseDrawArgs, "ctx">) {
-    const { ctx } = args;
-    ctx.textAlign = "start";
-}
-
-export function drawMarkerRowCell(
-    args: BaseDrawArgs,
-    index: number,
-    checked: boolean,
-    markerKind: "checkbox" | "both" | "number" | "checkbox-visible",
-    drawHandle: boolean
-) {
-    const { ctx, rect, hoverAmount, theme } = args;
-    const { x, y, width, height } = rect;
-    const checkedboxAlpha = checked ? 1 : markerKind === "checkbox-visible" ? 0.6 + 0.4 * hoverAmount : hoverAmount;
-    if (markerKind !== "number" && checkedboxAlpha > 0) {
-        ctx.globalAlpha = checkedboxAlpha;
-        const offsetAmount = 7 * (checked ? hoverAmount : 1);
-        drawCheckbox(
-            ctx,
-            theme,
-            checked,
-            drawHandle ? x + offsetAmount : x,
-            y,
-            drawHandle ? width - offsetAmount : width,
-            height,
-            true,
-            undefined,
-            undefined,
-            18
-        );
-        if (drawHandle) {
-            ctx.globalAlpha = hoverAmount;
-            ctx.beginPath();
-            for (const xOffset of [3, 6]) {
-                for (const yOffset of [-5, -1, 3]) {
-                    ctx.rect(x + xOffset, y + height / 2 + yOffset, 2, 2);
-                }
-            }
-
-            ctx.fillStyle = theme.textLight;
-            ctx.fill();
-            ctx.beginPath();
-        }
-        ctx.globalAlpha = 1;
-    }
-    if (markerKind === "number" || (markerKind === "both" && !checked)) {
-        const text = index.toString();
-
-        const start = x + width / 2;
-        if (markerKind === "both" && hoverAmount !== 0) {
-            ctx.globalAlpha = 1 - hoverAmount;
-        }
-        ctx.fillStyle = theme.textLight;
-        ctx.fillText(text, start, y + height / 2 + getMiddleCenterBias(ctx, `9px ${theme.fontFamily}`));
-        if (hoverAmount !== 0) {
-            ctx.globalAlpha = 1;
-        }
-    }
-}
-
-export function drawProtectedCell(args: BaseDrawArgs) {
-    const { ctx, theme, rect } = args;
-    const { x, y, height: h } = rect;
-
-    ctx.beginPath();
-
-    const radius = 2.5;
-    let xStart = x + theme.cellHorizontalPadding + radius;
-    const center = y + h / 2;
-    const p = Math.cos(degreesToRadians(30)) * radius;
-    const q = Math.sin(degreesToRadians(30)) * radius;
-
-    for (let i = 0; i < 12; i++) {
-        ctx.moveTo(xStart, center - radius);
-        ctx.lineTo(xStart, center + radius);
-
-        ctx.moveTo(xStart + p, center - q);
-        ctx.lineTo(xStart - p, center + q);
-
-        ctx.moveTo(xStart - p, center - q);
-        ctx.lineTo(xStart + p, center + q);
-        xStart += 8;
-    }
-    ctx.lineWidth = 1.1;
-    ctx.lineCap = "square";
-    ctx.strokeStyle = theme.textLight;
-    ctx.stroke();
-}
-
-interface CornerRadius {
-    tl: number;
-    tr: number;
-    bl: number;
-    br: number;
-}
-
-function roundedRect(
-    ctx: CanvasRenderingContext2D,
-    x: number,
-    y: number,
-    width: number,
-    height: number,
-    radius: number | CornerRadius
-) {
-    if (typeof radius === "number") {
-        radius = { tl: radius, tr: radius, br: radius, bl: radius };
-    }
-
-    // restrict radius to a reasonable max
-    radius = {
-        tl: Math.min(radius.tl, height / 2, width / 2),
-        tr: Math.min(radius.tr, height / 2, width / 2),
-        bl: Math.min(radius.bl, height / 2, width / 2),
-        br: Math.min(radius.br, height / 2, width / 2),
-    };
-
-    ctx.moveTo(x + radius.tl, y);
-    ctx.arcTo(x + width, y, x + width, y + radius.tr, radius.tr);
-    ctx.arcTo(x + width, y + height, x + width - radius.br, y + height, radius.br);
-    ctx.arcTo(x, y + height, x, y + height - radius.bl, radius.bl);
-    ctx.arcTo(x, y, x + radius.tl, y, radius.tl);
-}
-
-export function drawBoolean(
-    args: BaseDrawArgs,
-    data: boolean | BooleanEmpty | BooleanIndeterminate,
-    canEdit: boolean,
-    maxSize?: number
-) {
-    if (!canEdit && data === BooleanEmpty) {
-        return;
-    }
-
-    const { ctx, hoverAmount, theme, rect, highlighted, hoverX, hoverY } = args;
-    const { x, y, width: w, height: h } = rect;
-
-    const hoverEffect = 0.35;
-
-    let alpha = canEdit ? 1 - hoverEffect + hoverEffect * hoverAmount : 0.4;
-    if (data === BooleanEmpty) {
-        alpha *= hoverAmount;
-    }
-    if (alpha === 0) {
-        return;
-    }
-    ctx.globalAlpha = alpha;
-
-    drawCheckbox(ctx, theme, data, x, y, w, h, highlighted, hoverX, hoverY, maxSize);
-
-    ctx.globalAlpha = 1;
-}
-
-const itemMargin = 4;
-
-export function drawBubbles(args: BaseDrawArgs, data: readonly string[]) {
-    const { rect, theme, ctx, highlighted } = args;
-    const { x, y, width: w, height: h } = rect;
-    const bubbleHeight = 20;
-    const bubblePad = 8;
-    const bubbleMargin = itemMargin;
-    let renderX = x + theme.cellHorizontalPadding;
-
-    const renderBoxes: { x: number; width: number }[] = [];
-    for (const s of data) {
-        if (renderX > x + w) break;
-        const textWidth = measureTextCached(s, ctx, `${theme.baseFontStyle} ${theme.fontFamily}`).width;
-        renderBoxes.push({
-            x: renderX,
-            width: textWidth,
-        });
-
-        renderX += textWidth + bubblePad * 2 + bubbleMargin;
-    }
-
-    ctx.beginPath();
-    for (const rectInfo of renderBoxes) {
-        roundedRect(
-            ctx,
-            rectInfo.x,
-            y + (h - bubbleHeight) / 2,
-            rectInfo.width + bubblePad * 2,
-            bubbleHeight,
-            bubbleHeight / 2
-        );
-    }
-    ctx.fillStyle = highlighted ? theme.bgBubbleSelected : theme.bgBubble;
-    ctx.fill();
-
-    for (const [i, rectInfo] of renderBoxes.entries()) {
-        ctx.beginPath();
-        ctx.fillStyle = theme.textBubble;
-        ctx.fillText(data[i], rectInfo.x + bubblePad, y + h / 2 + getMiddleCenterBias(ctx, theme));
-    }
-}
-
-const drilldownCache: {
-    [key: string]: HTMLCanvasElement;
-} = {};
-
-function getAndCacheDrilldownBorder(
-    bgCell: string,
-    border: string,
-    height: number
-): {
-    el: HTMLCanvasElement;
-    height: number;
-    width: number;
-    middleWidth: number;
-    sideWidth: number;
-    dpr: number;
-    padding: number;
-} | null {
-    const dpr = Math.ceil(window.devicePixelRatio);
-    const shadowBlur = 5;
-    const targetHeight = height - shadowBlur * 2;
-    const middleWidth = 4;
-    const rounding = 6;
-
-    const innerHeight = height * dpr;
-    const sideWidth = rounding + shadowBlur;
-    const targetWidth = rounding * 3;
-    const innerWidth = (targetWidth + shadowBlur * 2) * dpr;
-
-    const key = `${bgCell},${border},${dpr},${height}`;
-    if (drilldownCache[key] !== undefined) {
-        return {
-            el: drilldownCache[key],
-            height: innerHeight,
-            width: innerWidth,
-            middleWidth: middleWidth * dpr,
-            sideWidth: sideWidth * dpr,
-            padding: shadowBlur * dpr,
-            dpr,
-        };
-    }
-
-    const canvas = document.createElement("canvas");
-    const ctx = canvas.getContext("2d"); // alpha needed
-
-    if (ctx === null) return null;
-
-    canvas.width = innerWidth;
-    canvas.height = innerHeight;
-
-    ctx.scale(dpr, dpr);
-
-    drilldownCache[key] = canvas;
-
-    const trueRounding = Math.min(rounding, targetWidth / 2, targetHeight / 2);
-    ctx.beginPath();
-    roundedRect(ctx, shadowBlur, shadowBlur, targetWidth, targetHeight, trueRounding);
-
-    ctx.shadowColor = "rgba(24, 25, 34, 0.4)";
-    ctx.shadowBlur = 1;
-    ctx.fillStyle = bgCell;
-    ctx.fill();
-
-    ctx.shadowColor = "rgba(24, 25, 34, 0.3)";
-    ctx.shadowOffsetY = 1;
-    ctx.shadowBlur = 5;
-    ctx.fillStyle = bgCell;
-    ctx.fill();
-
-    ctx.shadowOffsetY = 0;
-    ctx.shadowBlur = 0;
-    ctx.shadowBlur = 0;
-
-    ctx.beginPath();
-    roundedRect(ctx, shadowBlur + 0.5, shadowBlur + 0.5, targetWidth, targetHeight, trueRounding);
-
-    ctx.strokeStyle = border;
-    ctx.lineWidth = 1;
-    ctx.stroke();
-
-    return {
-        el: canvas,
-        height: innerHeight,
-        width: innerWidth,
-        sideWidth: sideWidth * dpr,
-        middleWidth: rounding * dpr,
-        padding: shadowBlur * dpr,
-        dpr,
-    };
-}
-
-export function drawDrilldownCell(args: BaseDrawArgs, data: readonly DrilldownCellData[]) {
-    const { rect, theme, ctx, imageLoader, col, row } = args;
-    const { x, width: w } = rect;
-
-    const font = `${theme.baseFontStyle} ${theme.fontFamily}`;
-    const emHeight = getEmHeight(ctx, font);
-    const h = Math.min(rect.height, Math.max(16, Math.ceil(emHeight * theme.lineHeight) * 2));
-    const y = Math.floor(rect.y + (rect.height - h) / 2);
-
-    const bubbleHeight = h - 10;
-    const bubblePad = 8;
-    const bubbleMargin = itemMargin;
-    let renderX = x + theme.cellHorizontalPadding;
-
-    const tileMap = getAndCacheDrilldownBorder(theme.bgCell, theme.drilldownBorder, h);
-
-    const renderBoxes: { x: number; width: number }[] = [];
-    for (const el of data) {
-        if (renderX > x + w) break;
-        const textMetrics = measureTextCached(el.text, ctx, font);
-        const textWidth = textMetrics.width;
-        let imgWidth = 0;
-        if (el.img !== undefined) {
-            const img = imageLoader.loadOrGetImage(el.img, col, row);
-            if (img !== undefined) {
-                imgWidth = bubbleHeight - 8 + 4;
-            }
-        }
-        const renderWidth = textWidth + imgWidth + bubblePad * 2;
-        renderBoxes.push({
-            x: renderX,
-            width: renderWidth,
-        });
-
-        renderX += renderWidth + bubbleMargin;
-    }
-
-    if (tileMap !== null) {
-        const { el, height, middleWidth, sideWidth, width, dpr, padding } = tileMap;
-        const outerSideWidth = sideWidth / dpr;
-        const outerPadding = padding / dpr;
-        for (const rectInfo of renderBoxes) {
-            const rx = Math.floor(rectInfo.x);
-            const rw = Math.floor(rectInfo.width);
-            ctx.imageSmoothingEnabled = false;
-
-            ctx.drawImage(el, 0, 0, sideWidth, height, rx - outerPadding, y, outerSideWidth, h);
-            if (rectInfo.width > sideWidth * 2)
-                ctx.drawImage(
-                    el,
-                    sideWidth,
-                    0,
-                    middleWidth,
-                    height,
-                    rx + (outerSideWidth - outerPadding),
-                    y,
-                    rw - (outerSideWidth - outerPadding) * 2,
-                    h
-                );
-            ctx.drawImage(
-                el,
-                width - sideWidth,
-                0,
-                sideWidth,
-                height,
-                rx + rw - (outerSideWidth - outerPadding),
-                y,
-                outerSideWidth,
-                h
-            );
-            ctx.imageSmoothingEnabled = true;
-        }
-    }
-
-    ctx.beginPath();
-
-    for (const [i, rectInfo] of renderBoxes.entries()) {
-        const d = data[i];
-        let drawX = rectInfo.x + bubblePad;
-
-        if (d.img !== undefined) {
-            const img = imageLoader.loadOrGetImage(d.img, col, row);
-            if (img !== undefined) {
-                const imgSize = bubbleHeight - 8;
-                let srcX = 0;
-                let srcY = 0;
-                let srcWidth = img.width;
-                let srcHeight = img.height;
-
-                if (srcWidth > srcHeight) {
-                    // landscape
-                    srcX += (srcWidth - srcHeight) / 2;
-                    srcWidth = srcHeight;
-                } else if (srcHeight > srcWidth) {
-                    //portrait
-                    srcY += (srcHeight - srcWidth) / 2;
-                    srcHeight = srcWidth;
-                }
-                ctx.beginPath();
-                roundedRect(ctx, drawX, y + h / 2 - imgSize / 2, imgSize, imgSize, 3);
-                ctx.save();
-                ctx.clip();
-                ctx.drawImage(img, srcX, srcY, srcWidth, srcHeight, drawX, y + h / 2 - imgSize / 2, imgSize, imgSize);
-                ctx.restore();
-
-                drawX += imgSize + 4;
-            }
-        }
-
-        ctx.beginPath();
-        ctx.fillStyle = theme.textBubble;
-        ctx.fillText(d.text, drawX, y + h / 2 + getMiddleCenterBias(ctx, theme));
-    }
-}
-
-export function drawImage(
-    args: BaseDrawArgs,
-    data: readonly string[],
-    rounding: number = 4,
-    contentAlign?: BaseGridCell["contentAlign"]
-) {
-    const { rect, col, row, theme, ctx, imageLoader } = args;
-    const { x, y, height: h, width: w } = rect;
-
-    const imgHeight = h - theme.cellVerticalPadding * 2;
-    const images: (HTMLImageElement | ImageBitmap)[] = [];
-    let totalWidth = 0;
-    for (let index = 0; index < data.length; index++) {
-        const i = data[index];
-        if (i.length === 0) continue;
-        const img = imageLoader.loadOrGetImage(i, col, row);
-
-        if (img !== undefined) {
-            images[index] = img;
-            const imgWidth = img.width * (imgHeight / img.height);
-            totalWidth += imgWidth + itemMargin;
-        }
-    }
-
-    if (totalWidth === 0) return;
-    totalWidth -= itemMargin;
-
-    let drawX = x + theme.cellHorizontalPadding;
-    if (contentAlign === "right") drawX = Math.floor(x + w - theme.cellHorizontalPadding - totalWidth);
-    else if (contentAlign === "center") drawX = Math.floor(x + w / 2 - totalWidth / 2);
-
-    for (const img of images) {
-        if (img === undefined) continue; //array is sparse
-        const imgWidth = img.width * (imgHeight / img.height);
-        if (rounding > 0) {
-            roundedRect(ctx, drawX, y + theme.cellVerticalPadding, imgWidth, imgHeight, rounding);
-            ctx.save();
-            ctx.clip();
-        }
-        ctx.drawImage(img, drawX, y + theme.cellVerticalPadding, imgWidth, imgHeight);
-        if (rounding > 0) {
-            ctx.restore();
-        }
-
-        drawX += imgWidth + itemMargin;
     }
 }
