+++ packages/core/src/data-grid/data-grid-sprites.ts
@@ -1,6 +1,6 @@
 import type { Theme } from "../common/styles";
 import type { SpriteProps } from "../common/utils";
-import { HeaderIconMap, sprites } from "./sprites";
+import type { HeaderIconMap } from "./sprites";
 
 /**
  * A known icon identifier
@@ -46,7 +46,6 @@ export class SpriteManager {
 
     constructor(headerIcons: SpriteMap | undefined, private onSettled: () => void) {
         this.headerIcons = {
-            ...sprites,
             ...headerIcons,
         };
     }
