+++ packages/core/src/data-grid/data-grid.stories.tsx
@@ -32,7 +32,7 @@ const emptyGridSelection: GridSelection = {
 
 const getCellRenderer: GetCellRendererCallback = cell => {
     if (cell.kind === GridCellKind.Custom) return undefined;
-    return CellRenderers[cell.kind] as any;
+    return CellRenderers.find(x => x.kind === cell.kind) as any;
 };
 
 export function Simplenotest() {
