+++ packages/core/src/scrolling-data-grid/scrolling-data-grid.stories.tsx
@@ -21,7 +21,7 @@ const InnerContainer = styled.div`
 
 const getCellRenderer: GetCellRendererCallback = cell => {
     if (cell.kind === GridCellKind.Custom) return undefined;
-    return CellRenderers[cell.kind] as any;
+    return CellRenderers.find(x => x.kind === cell.kind) as any;
 };
 
 export default {
