+++ packages/core/test/use-column-sizer.test.tsx
@@ -79,7 +79,7 @@ const abortController = new AbortController();
 
 const getCellRenderer: GetCellRendererCallback = cell => {
     if (cell.kind === GridCellKind.Custom) return undefined;
-    return CellRenderers[cell.kind] as any;
+    return CellRenderers.find(x => x.kind === cell.kind) as any;
 };
 
 describe("use-column-sizer", () => {
